split types and js into separate build files
