+++ package.json
@@ -6,6 +6,7 @@
   "engines": {
     "node": ">=20.8.1"
   },
+  "sideEffects": false,
   "publishConfig": {
     "access": "public",
     "provenance": true
@@ -57,8 +58,16 @@
     "webpack-dev-server": "^5.2.0",
     "zigar-loader": "^0.14.1"
   },
-  "types": "./lib/types/index.d.ts",
-  "main": "./lib/index.mjs",
+  "exports": {
+    ".": {
+      "types": "./lib/types/index.d.ts",
+      "import": "./lib/index.mjs"
+    },
+    "./types": {
+      "types": "./lib/types/types.d.ts",
+      "import": "./lib/types.mjs"
+    }
+  },
   "type": "module",
   "repository": {
     "type": "git",
